### 2017-11-10 #452 @arfon
👋 @bmagyar. Thanks for your submission to JOSS. A couple of questions/clarifications on this submission:

- Could you confirm what the research application of the `ros-controls` suite of software is?
- I'm concerned about the amount of work any reviewer is going to have to do here to review this submission. We generally expect one repository per submission but this looks to be a collection of separate software repositories. Is there a 'main' repository that it might make more sense to submit?
