### 2017-11-10 #452 @bmagyar
Hi @arfon

`ros_control` is a framework for robot control for robots using the Robot Operating System (ROS). ROS is core to most modern robotics research and it is making it's way to industry as well (e.g.: warehouse robots, home-care research platforms, commercial & research-grade self driving cars). The last section of our paper has a list of some research robots where `ros_control` is used to provide the control architecture & allows for researchers to create their custom controllers using the framework. This project is rather mature, hence the size and number of applications.

I was indeed a bit worried that the review may get a wee bit tricky...

We split the framework among different repositories to ease the binary release process as some packages such as message definitions and core libraries are changed less frequently than our off-the-shelf controllers. I don't think it'd make sense to assess only one part of it.

I can try to help by providing a more detailed guide with pointing out all sources of documentation, how to set things up on one's machine for a quick test and pointers to our continuous integration & test suite with Travis.
