### 2017-11-10 #452 @arfon
> We split the framework among different repositories to ease the binary release process as some packages such as message definitions and core libraries are changed less frequently than our off-the-shelf controllers. I don't think it'd make sense to assess only one part of it.

OK thanks @bmagyar.

> I can try to help by providing a more detailed guide with pointing out all sources of documentation, how to set things up on one's machine for a quick test and pointers to our continuous integration & test suite with Travis.

Yes, I think that might be useful. Also, a forewarning: we might struggle a little to find an editor and reviewers for this submission as it's somewhat outside our 'regular' subject matter.
