### 2017-11-10 #452 @danielskatz
I'm a little familiar with ROS, and could act as editor, given good suggestions for reviewers.
