### 2017-11-11 #452 @arfon
> Am I allowed to suggest or does it have to come from the editors?

We'd welcome some names (and ideally GitHub handles) for some potential reviewers.
