### 2017-11-14 #452 @danielskatz
@bmagyar, please offer some suggestions (that don't have conflicts, of course)
