### 2017-11-14 #452 @bmagyar
* Jose Luis Rivero: worked for the Open Source Robotics Foundation (creators & maintainers of ROS) for years | https://github.com/j-rivero

* Aris Synodinos: University of Patras, Robotics Group  | https://github.com/progtologist

* Miguel Prada: works for Tecnalia, Spain. | https://github.com/miguelprada

There are 3 more people on my list, waiting for confirmation of interest.
