### 2017-11-14 #452 @bmagyar
* Jorge Nicho (will be away but can either volunteer a colleague or review when back): works at Southwest Research Institute, they are part of the ROS-Industrial consortium  | https://github.com/jrgnicho
