### 2017-11-14 #452 @danielskatz
@j-rivero, @progtologist, @miguelprada - would you be willing to review this submission for JOSS?

Please see http://joss.theoj.org/about#reviewer_guidelines for reviewer guidelines.

Also, please be sure you do not have any conflict of interest.
