### 2017-11-15 #456 @whedon
[REVIEW]: ros_control: A generic and simple control framework for ROS
**Submitting author:** @bmagyar (<a href="http://orcid.org/0000-0001-8455-8674">Bence Magyar</a>)
**Repository:** <a href="https://github.com/ros-controls/joss_paper" target ="_blank">https://github.com/ros-controls/joss_paper</a>
**Version:** 0.12.0
