### 2017-11-15 #456 @miguelprada
@danielskatz My two first questions already:

First, I might be missing something really obvious here, but I cannot seem to mark the checklist items in the original comment.

Second, the repository listed in the original comment (both in the header and the first item in _general checks_) is to the contents of the paper, not the software itself. Is this ok?

To be clear, I know where to find the software, I just hesitated when I saw the text
> Repository: Is the source code for this software available at the repository url?

with a link pointing to a repo containing just the paper.
