### 2017-11-15 #456 @danielskatz
@migueldvb sorry, I need to assign you to this issue, and am working on that.

regarding the second point, you are correct, and the repo should point up one level, not to the paper.

@arfon, help on both of these points is welcome.
