### 2017-11-15 #456 @arfon
> First, I might be missing something really obvious here, but I cannot seem to mark the checklist items in the original comment.

@miguelprada - you should have access now (please accept the invite from the openjournals organization).

> regarding the second point, you are correct, and the repo should point up one level, not to the paper.

@bmagyar - could you please advise?
