### 2017-11-16 #456 @miguelprada
> > First, I might be missing something really obvious here, but I cannot seem to mark the checklist items in the original comment.

> @miguelprada - you should have access now (please accept the invite from the openjournals organization).

Accepted the org invite, but I'm still unable to check the items... :sweat:

May I be doing something wrong? It should be just a matter of clicking on the checkboxes, right? I don't see any other way to edit the original post either.
