### 2017-11-16 #456 @arfon
> Accepted the org invite, but I'm still unable to check the items... 😓
Sorry, one more thing. Please accept the invite on this page: https://github.com/openjournals/joss-reviews/invitations

(This usually happens automatically via the API but as we can't handle multiple reviewers automatically yet I have to give you access manually)
