### 2017-11-17 #452 @j-rivero
> @j-rivero, @progtologist, @miguelprada - would you be willing to review this submission for JOSS?

Sorry for not answering before, I'm happy to help if you sill need someone to perform a review.
