### 2017-11-22 ros-controls/joss_paper#23 @progtologist
LICENSE
JOSS requires a LICENSE file to be provided in the repository. Strictly speaking this is not the case with ros_control, though the license is clearly stated inside each source file.

I am not sure whether this should be "fixed" or be considered redundant.
