### 2017-11-23 ros-controls/joss_paper#23 @bmagyar
We also specify the license at the package level in each `package.xml`.

It is probably best to refer to the editors.
