### 2017-11-23 #456 @progtologist
@danielskatz can you provide your feedback regarding [this](https://github.com/ros-controls/joss_paper/issues/23#issuecomment-346631876) issue?
