### 2017-11-23 ros-controls/joss_paper#23 @danielskatz
JOSS requires a ```LICENSE``` or ```LICENSE.txt``` file in the root of the repository.
