### 2017-11-23 #456 @danielskatz
As I just wrote in the issue:

JOSS requires a ```LICENSE``` or ```LICENSE.txt``` file in the root of the repository.
