### 2017-11-23 ros-controls/joss_paper#23 @bmagyar
Done:
https://github.com/ros-controls/ros_control/commit/d8a2c5e805a22b6f53c48789bc5dd47692e47772
