### 2017-11-24 #456 @progtologist
Grade: 1) Accept

All minor issues are now resolved as far as I am concerned. The software is research orientated with wide acceptance in the field and is actively developed and maintained.
