### 2017-11-27 ros-controls/joss_paper#24 @miguelprada
References should list DOI where applicable
Some of the references that should list a DOI, as required by JOSS, are missing it.
