### 2017-11-27 #456 @miguelprada
I'll issue a grade 2, if only for the [missing DOI entries](https://github.com/ros-controls/joss_paper/issues/24). Will replace it with a 1 once they're in place.
