+++ paper.bib
@@ -1,47 +1,7 @@
 % This file was created with JabRef 2.10.
 % Encoding: UTF-8
 
-
-@InProceedings{jorg2014hardware,
-  Title                    = {The hardware abstraction layer—supporting control design by tackling the complexity of humanoid robot hardware},
-  Author                   = {J{\"o}rg, Stefan and Tully, Jan and Albu-Sch{\"a}ffer, Alin},
-  Booktitle                = {Robotics and Automation (ICRA), 2014 IEEE International Conference on},
-  Year                     = {2014},
-  Organization             = {IEEE},
-  Pages                    = {6427--6433},
-
-  File                     = {:pdfs/jorg2014hardware.pdf:PDF}
-}
-
-@Misc{murray2005hardware,
-  Title                    = {Hardware abstraction layer for a robot},
-
-  Author                   = {Murray IV, Thomas J and Pham, Baoquoc N and Pirjanian, Paolo},
-  Month                    = may # {~3},
-  Note                     = {US Patent 6,889,118},
-  Year                     = {2005},
-
-  Publisher                = {Google Patents}
-}
-
-@Article{schaal2009sl,
-  Title                    = {The SL simulation and real-time control software package},
-  Author                   = {Schaal, Stefan},
-  Journal                  = {University of Southern California},
-  Year                     = {2009}
-}
-
 % PUT ENABLED PUBLICATIONS HERE
-
-@inproceedings{bruyninckx2001open,
-  title={Open robot control software: the OROCOS project},
-  author={Bruyninckx, Herman},
-  booktitle={Robotics and Automation, 2001. Proceedings 2001 ICRA. IEEE International Conference on},
-  volume={3},
-  pages={2523--2528},
-  year={2001},
-  organization={IEEE}
-}
 
 @article{chitta2012moveit,
   title={Moveit![ROS topics]},
@@ -54,41 +14,14 @@
   publisher={IEEE}
 }
 
-@InProceedings{muratore2017xbotcore,
-  Title                    = {XBotCore: A Real-Time Cross-Robot Software Platform},
-  Author                   = {Muratore, Luca and Laurenzi, Arturo and Hoffman, Enrico Mingo and Rocchi, Alessio and Caldwell, Darwin G and Tsagarakis, Nikos G},
-  Booktitle                = {Robotic Computing (IRC), IEEE International Conference on},
-  Year                     = {2017},
-  Organization             = {IEEE},
-  Pages                    = {77--80}
-}
-
 @incollection{badger2016ros,
   title={ROS in Space: A Case Study on Robonaut 2},
   author={Badger, Julia and Gooding, Dustin and Ensley, Kody and Hambuchen, Kimberly and Thackston, Allison},
   booktitle={Robot Operating System (ROS)},
   pages={343--373},
   year={2016},
+  doi={10.1007/978-3-319-26054-9_13},
   publisher={Springer}
-}
-
-
-@inproceedings{barros2014mobile,
-  title={A mobile manipulator controller implemented in the robot operating system},
-  author={Barros, Taiser Tadeu Teixeira and Lages, Walter Fetter},
-  booktitle={ISR/Robotik 2014; 41st International Symposium on Robotics; Proceedings of},
-  pages={1--8},
-  year={2014},
-  organization={VDE}
-}
-
-@inproceedings{romano2017whole,
-  title={A whole-body software abstraction layer for control design of free-floating mechanical systems},
[…]
