### 2017-11-27 4cfd39e
Update paper.bib
