+++ paper.bib
@@ -102,7 +102,7 @@
 }
 
 @techreport{andersen2015optimizing,
-  title={Optimizing the Universal Robots ROS driver},
+  title={Optimizing the Universal Robots ROS driver.},
   author={Andersen, Thomas Timm},
   year={2015},
   institution={Technical University of Denmark, Department of Electrical Engineering}
