### 2017-11-27 ros-controls/joss_paper#24 @bmagyar
Done!
