### 2017-11-27 #456 @arfon
@bmagyar - At this point could you make an archive of the reviewed software packages in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
