### 2017-11-28 #456 @bmagyar
What would you recommend for gathering up all the package into a tar and submitting for doi?
Zenodo only seems to support repo-wise archiving, figshare seems to be ok with the big tarball.

Should I add a pdf generated by me to the joss-paper repo or will that be generated by them/you?
