### 2017-11-30 #456 @arfon
> What would you recommend for gathering up all the package into a tar and submitting for doi?
Zenodo only seems to support repo-wise archiving, figshare seems to be ok with the big tarball.

My understanding is that both figshare and Zenodo will let you archive a bunch of zip/tar files with them, Zenodo just has a cleaner/more automated approach if you're archiving a single repository.

> Should I add a pdf generated by me to the joss-paper repo or will that be generated by them/you?

That will be done by us. We just just need you to supply us with the archive DOI for the software that has been reviewed as part of this submission.
