### 2017-12-04 c663c2c
attempt fix of Jonathan Bohren affil to fix PDF comp. error
adding a space between the affiliations
