### 2017-12-04 #456 @danielskatz
@bmagyar , can you merge https://github.com/ros-controls/joss_paper/pull/26 please?
Then we'll see if this fixes the problem generating the PDF
