### 2017-12-04 #456 @bmagyar
Thanks for giving a shot to fix it. It's merged
