### 2017-12-04 #456 @danielskatz
hey @arfon - this is a very painful way to debug markdown/bibtex - PDF.  Maybe we need a locally runnable option as well?
