### 2017-12-04 #456 @danielskatz
@bmagyar - while the reviewers checked the references box, it looks like there is at least one paper in the bibtex file, the last one (radford2015valkyrie), that has a DOI but where the DOI is not part of the bibtext.

Please make sure all articles that you include in the bibtex that have DOIs have them in their bibtex entries.
