### 2017-12-04 8c3c718
Update Valkyrie reference for one with DOI
