+++ paper.bib
@@ -123,13 +123,17 @@
   year={2009}
 }
 
-@article{radford2015valkyrie,
-  title={Valkyrie: NASA's first bipedal humanoid robot},
-  author={Radford, Nicolaus A and Strawser, Philip and Hambuchen, Kimberly and Mehling, Joshua S and Verdeyen, William K and Donnan, A Stuart and Holley, James and Sanchez, Jairo and Nguyen, Vienny and Bridgwater, Lyndon and others},
-  journal={Journal of Field Robotics},
-  volume={32},
-  number={3},
-  pages={397--419},
-  year={2015},
-  publisher={Wiley Online Library}
+
+
+@article {ROB:ROB21560,
+  author = {Radford, Nicolaus A. and Strawser, Philip and Hambuchen, Kimberly and Mehling, Joshua S. and Verdeyen, William K. and Donnan, A. Stuart and Holley, James and Sanchez, Jairo and Nguyen, Vienny and Bridgwater, Lyndon and Berka, Reginald and Ambrose, Robert and Myles Markee, Mason and Fraser-Chanpong, N. J. and McQuin, Christopher and Yamokoski, John D. and Hart, Stephen and Guo, Raymond and Parsons, Adam and Wightman, Brian and Dinh, Paul and Ames, Barrett and Blakely, Charles and Edmondson, Courtney and Sommers, Brett and Rea, Rochelle and Tobler, Chad and Bibby, Heather and Howard, Brice and Niu, Lei and Lee, Andrew and Conover, Michael and Truong, Lily and Reed, Ryan and Chesney, David and Platt, Robert and Johnson, Gwendolyn and Fok, Chien-Liang and Paine, Nicholas and Sentis, Luis and Cousineau, Eric and Sinnet, Ryan and Lack, Jordan and Powell, Matthew and Morris, Benjamin and Ames, Aaron and Akinyode, Jide},
+  title = {Valkyrie: NASA's First Bipedal Humanoid Robot},
+  journal = {Journal of Field Robotics},
+  volume = {32},
+  number = {3},
+  issn = {1556-4967},
+  url = {http://dx.doi.org/10.1002/rob.21560},
+  doi = {10.1002/rob.21560},
+  pages = {397--419},
+  year = {2015},
 }
