+++ paper.md
@@ -103,7 +103,7 @@ Being a mature framework, `ros_control` is widely applied to both production and
 - Clearpath Robotics' outdoor mobile robots: Grizzly, Husky, Jackal [@cpr2017roscontrol], and OTTO Motors' industrial indoor mobile robots: OTTO 1500, OTTO 100
 - The "Twil" robot at Federal University of Rio Grande do Sul [@lages2017parametric]
 - The quadruped robots HyQ and HyQ2Max [@semini11hyqdesign, @semini2017design] at Istituto Italiano di Tecnologia
-- NASA's humanoid and biped robots: Valkyrie & Robonaut [@radford2015valkyrie, @hart2014robot, @badger2016ros]
+- NASA's humanoid and biped robots: Valkyrie & Robonaut [@ROB:ROB21560, @hart2014robot, @badger2016ros]
 - PAL Robotics' humanoid, biped and mobile robots: REEM, REEM-C, PMB2, Tiago and Talos [@stasse2017talos] 
 - Shadow Robot's anthropomorphic, highly sensorized and precise Shadow Hand [@meier2016distinguishing]
 - Universal Robots' industrial arms: UR3, UR5 [@andersen2015optimizing]
