### 2017-12-04 #456 @bmagyar
Sorry for the trouble, indeed it would be nice to have something I could run locally.

I've updated the bibtex with one I took from Wiley directly, they've got DOI and ISSN.
