### 2017-12-04 #456 @arfon
> hey @arfon - this is a very painful way to debug markdown/bibtex -> PDF. Maybe we need a locally runnable option as well?

Yes, there is a way to run locally. You need to download and install the [Whedon gem](https://github.com/openjournals/whedon) (or Pandoc etc locally).

In this case, the error is due to missing commas in the bibtex file which I've fixed up here: https://github.com/ros-controls/joss_paper/pull/27
