### 2017-12-04 492dd25
Fixing typos in bibtex (#27)
