+++ paper.bib
@@ -47,7 +47,7 @@
   publisher={IEEE}
 }
 
-@article{semini2011design,
+@article{semini11hyqdesign,
   title={Design of HyQ--a hydraulically and electrically actuated quadruped robot},
   author={Semini, Claudio and Tsagarakis, Nikos G and Guglielmino, Emanuele and Focchi, Michele and Cannella, Ferdinando and Caldwell, Darwin G},
   journal={Proceedings of the Institution of Mechanical Engineers, Part I: Journal of Systems and Control Engineering},
@@ -77,7 +77,7 @@
 
 @misc{cpr2017roscontrol,
   title = {New Universal Robots Driver Makes Manipulation Research Easier},
-  howpublished = {\url{https://www.clearpathrobotics.com/2016/02/new-universal-robots-driver-makes-manipulation-easier}}
+  howpublished = {\url{https://www.clearpathrobotics.com/2016/02/new-universal-robots-driver-makes-manipulation-easier}},
   note = {Accessed: 2017-10-31}
 }
 
@@ -112,7 +112,7 @@
 @misc{rodriguez2014roscon,
   author = {{Adolfo Rodríguez Tsouroukdissian}},
   title = {[ROSCon2014] ros_control: An overview},
-  howpublished = {\url{https://vimeo.com/107507546}}
+  howpublished = {\url{https://vimeo.com/107507546}},
   note = {Accessed: 2017-10-31}
 }
 
