### 2017-12-04 #456 @arfon
> - In the PDF, the "repository" link on the left points to the paper repo (https://github.com/ros-controls/joss_paper) rather than the high-level set of repos (https://github.com/ros-controls) I understand why this is the case, but in this unusual paper describing a set of repos, this link should probably be changed manually.

OK, I can fix that up.

> Also in the PDF, the archive link seems wrong - it seems to point to the code repo, not to the zenodo archive...

I'm not sure it currently links to *anything*. This link is only made in the final version of the paper.

Assuming both of the above are fixed up, is this paper good to accept? If so, I can do the final pieces.
