### 2017-12-04 07e2175
misc bib fixes - also noting URL not found
http://www.ros.org/urdf/ doesn't exist - please fix it to what it should be.
