### 2017-12-04 #456 @danielskatz
let's wait until https://github.com/ros-controls/joss_paper/pull/28 is merged, which fixes some bib issues and also includes a note about a correct URL that @bmagyar needs to find (as I can't find the right one)
