+++ paper.bib
@@ -122,7 +122,7 @@
 @article{garage2009universal,
   title={Universal Robot Description Format ({URDF})},
   author={{Willow Garage}},
-  howpublished={\url{http://www.ros.org/urdf/}},
+  howpublished={\url{http://wiki.ros.org/urdf/}},
   year={2009}
 }
 
