### 2017-12-04 7e56757
misc bib fixes - also noting URL not found (#28)
* misc bib fixes - also noting URL not found

http://www.ros.org/urdf/ doesn't exist - please fix it to what it should be.

* fixing the bad URL
