+++ paper.bib
@@ -15,9 +15,9 @@
 }
 
 @incollection{badger2016ros,
-  title={ROS in Space: A Case Study on Robonaut 2},
+  title={{ROS} in Space: A Case Study on Robonaut 2},
   author={Badger, Julia and Gooding, Dustin and Ensley, Kody and Hambuchen, Kimberly and Thackston, Allison},
-  booktitle={Robot Operating System (ROS)},
+  booktitle={Robot Operating System ({ROS})},
   pages={343--373},
   year={2016},
   doi={10.1007/978-3-319-26054-9_13},
@@ -25,18 +25,19 @@
 }
 
 @inproceedings{quigley2009ros,
-  title={ROS: an open-source Robot Operating System},
+  title={{ROS}: an open-source Robot Operating System},
   author={Quigley, Morgan and Conley, Ken and Gerkey, Brian and Faust, Josh and Foote, Tully and Leibs, Jeremy and Wheeler, Rob and Ng, Andrew Y},
-  booktitle={ICRA workshop on open source software},
+  booktitle={ICRA Workshop on Open Source Software},
   volume={3},
   number={3.2},
   pages={5},
   year={2009},
-  organization={Kobe}
+  organization={Kobe},
+  howpublished={\url{http://www.willowgarage.com/papers/ros-open-source-robot-operating-system}}
 }
 
 @article{semini2017design,
-  title={Design of the Hydraulically Actuated, Torque-Controlled Quadruped Robot HyQ2Max},
+  title={Design of the Hydraulically Actuated, Torque-Controlled Quadruped Robot {HyQ2Max}},
   author={Semini, Claudio and Barasuol, Victor and Goldsmith, Jake and Frigerio, Marco and Focchi, Michele and Gao, Yifu and Caldwell, Darwin G},
   journal={IEEE/ASME Transactions on Mechatronics},
   volume={22},
@@ -48,7 +49,7 @@
 }
 
 @article{semini11hyqdesign,
-  title={Design of HyQ--a hydraulically and electrically actuated quadruped robot},
+  title={Design of {HyQ}--a hydraulically and electrically actuated quadruped robot},
   author={Semini, Claudio and Tsagarakis, Nikos G and Guglielmino, Emanuele and Focchi, Michele and Cannella, Ferdinando and Caldwell, Darwin G},
   journal={Proceedings of the Institution of Mechanical Engineers, Part I: Journal of Systems and Control Engineering},
   volume={225},
@@ -60,15 +61,16 @@
 }
 
 @article{stasse2017talos,
-  title={TALOS: A new humanoid research platform targeted for industrial applications},
+  title={{TALOS}: A new humanoid research platform targeted for industrial applications},
   author={Stasse, Olivier and Flayols, Thomas and Budhiraja, Rohan and Giraud-Esclasse, Kevin and Carpentier, Justin and Del Prete, Andrea and Soueres, Philippe and Mansard, Nicolas and Lamiraux, Florent and Laumond, Jean-Paul and others},
+  howpublished={\url{https://hal.archives-ouvertes.fr/hal-01485519}},
   year={2017}
 }
 
 @inproceedings{hart2014robot,
-  title={Robot Task Commander: A framework and IDE for robot application development},
+  title={Robot Task Commander: A framework and {IDE} for robot application development},
   author={Hart, Stephen and Dinh, Paul and Yamokoski, John D and Wightman, Brian and Radford, Nicolaus},
-  booktitle={Intelligent Robots and Systems (IROS 2014), 2014 IEEE/RSJ International Conference on},
[…]
