### 2017-12-04 #456 @danielskatz
@arfon - any idea what got stuck here?  Reticulation of splines is usually such a quick operation...
