### 2017-12-04 #456 @arfon
> @arfon - any idea what got stuck here? Reticulation of splines is usually such a quick operation...

I think Heroku might be doing some platform maintenance right now (I'm seeing some weird behaviour elsewhere). I'll do some digging.
