### 2017-12-05 #456 @graiola
Hello, I am @graiola, one of the authors.
First of all, I would like to thank you all for your efforts in making this publication possible.
By looking at the resulting pdf I noticed that  the markdown list has not been correctly transformed (look at the robot section). Is it possible to fix it? Thanks in advance.
