### 2017-12-05 #456 @danielskatz
@arfon - if you look at https://github.com/ros-controls/joss_paper/blob/master/paper.md you will see a section
# Robots using ```ros_control```
that has a list of bullets.  In both the version at the DOI and the PDF, "ros_control" is in a too-small font, and the bullets have been lost.

Can you fix these?
