### 2017-12-05 #456 @arfon
Looks like Pandoc is really fussy about whitespace around the list. I've fixed it up in https://github.com/ros-controls/joss_paper/pull/29
