### 2017-12-05 #456 @arfon
> @arfon - The PR has been merged - can you regenerate the paper now?

Yes, I've regenerated it.
