### 2017-12-05 #456 @danielskatz
both the version at the DOI and the PDF still look not correct to me.  Am I missing something? Maybe I'm getting a cached version?
