### 2017-12-05 #456 @arfon
![screen shot 2017-12-05 at 10 45 26 am](https://user-images.githubusercontent.com/4483/33615956-754f0f42-d9a9-11e7-887c-1707cfccd172.png)

![screen shot 2017-12-05 at 11 03 05 am](https://user-images.githubusercontent.com/4483/33616772-e6e32f92-d9ab-11e7-9d53-bf132db26e34.png)

Yep, I think you must be getting a cached versions? See the screenshot of the PDF view and the website.
